Extract skill section into new file
